Revert "refactor: remodel help modal (#322)"
This reverts commit c6a473f08b39c93e0757c893c16d027f922ec70d.
