+++ src/ui/css/styles.css
@@ -598,61 +598,3 @@ textarea {
 .settings-row-input {
   width: 100%;
 }
-
-/* Help Dialog Styles */
-.help-dialog {
-  z-index: 9998;
-}
-
-.help-search {
-  margin-bottom: 1rem;
-}
-
-.help-group-list {
-  margin-top: 0.2rem;
-  padding-left: 0.5rem;
-  padding-right: 0.5rem;
-}
-
-.help-group-header {
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  font-weight: bold;
-  padding: 0.5rem 0;
-}
-
-.help-group-title {
-  font-weight: bold;
-  margin-left: 0.5rem;
-}
-
-.help-row {
-  padding: 0.5rem 0;
-  margin-left: 0.5rem;
-}
-
-.help-no-results {
-  text-align: center;
-  padding: 2rem 0;
-  color: var(--text-secondary);
-}
-
-#search-description {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
-
-/* Responsive adjustments */
-@media (max-width: 600px) {
-  .help-row {
-    flex-direction: column;
-  }
-}
