+++ src/ui/pages/Help.tsx
@@ -1,174 +1,85 @@
-import type { HelpMenuItem } from '@type/help';
 import {
   Button,
   Dialog,
   DialogActions,
   DialogContent,
   DialogTitle,
+  Divider,
   Grid,
   Typography,
 } from '@mui/material';
 import { useViewModel } from '@state/hook/useViewModel';
-import { HelpGroup } from '@ui/components/help/HelpGroup';
-import { HelpSearch } from '@ui/components/help/HelpSearch';
-import { HELP_GROUP_FILTERS, HELP_GROUPS } from '@ui/react_constants';
-import React, { useId, useMemo, useState } from 'react';
+import React, { useId } from 'react';
 
-function filterHelpItems(items: HelpMenuItem[], filterFn: ((item: HelpMenuItem) => boolean) | ((item: HelpMenuItem, otherGroups: HelpMenuItem[][]) => boolean), otherGroups?: HelpMenuItem[][]): HelpMenuItem[] {
-  if (otherGroups) {
-    return items.filter(item => (filterFn as (item: HelpMenuItem, otherGroups: HelpMenuItem[][]) => boolean)(item, otherGroups));
-  }
-  return items.filter(item => (filterFn as (item: HelpMenuItem) => boolean)(item));
+interface HelpRowProps {
+  label: string;
+  shortcut: string;
 }
+
+const HelpRow: React.FC<HelpRowProps> = ({ label, shortcut }) => (
+  <Grid
+    container
+    spacing={1}
+    alignItems="center"
+    sx={{ py: 1 }}
+  >
+    <Grid size={{ xs: 12, sm: 6, md: 7 }}>
+      <Typography variant="body2">
+        {label}
+      </Typography>
+    </Grid>
+    <Grid size={{ xs: 12, sm: 6, md: 5 }}>
+      <Typography variant="body2" fontWeight={300}>
+        {shortcut}
+      </Typography>
+    </Grid>
+  </Grid>
+);
 
 const Help: React.FC = () => {
   const id = useId();
-  const titleId = useId();
-  const descriptionId = useId();
   const viewModel = useViewModel('help');
   const { items } = viewModel.state;
-  const [expandedGroup, setExpandedGroup] = useState<string | null>(null);
-  const [searchQuery, setSearchQuery] = useState('');
-
-  const navigationItems = useMemo(() =>
-    filterHelpItems(items, HELP_GROUP_FILTERS.Navigation), [items]);
-
-  const modeItems = useMemo(() =>
-    filterHelpItems(items, HELP_GROUP_FILTERS.Modes), [items]);
-
-  const autoplayItems = useMemo(() =>
-    filterHelpItems(items, HELP_GROUP_FILTERS['Autoplay Controls']), [items]);
-
-  const labelItems = useMemo(() =>
-    filterHelpItems(items, HELP_GROUP_FILTERS['Label Announcements']), [items]);
-
-  const generalItems = useMemo(() =>
-    filterHelpItems(
-      items,
-      HELP_GROUP_FILTERS['General Controls'],
-      [navigationItems, modeItems, autoplayItems, labelItems],
-    ), [items, navigationItems, modeItems, autoplayItems, labelItems]);
-
-  const getGroupItems = useMemo(() => (groupTitle: string): HelpMenuItem[] => {
-    switch (groupTitle) {
-      case 'Navigation':
-        return navigationItems;
-      case 'Modes':
-        return modeItems;
-      case 'Autoplay Controls':
-        return autoplayItems;
-      case 'Label Announcements':
-        return labelItems;
-      case 'General Controls':
-        return generalItems;
-      default:
-        return [];
-    }
-  }, [navigationItems, modeItems, autoplayItems, labelItems, generalItems]);
 
   const handleClose = (): void => {
     viewModel.toggle();
   };
 
-  const handleGroupToggle = (groupTitle: string): void => {
-    setExpandedGroup(expandedGroup === groupTitle ? null : groupTitle);
-  };
-
-  const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>): void => {
-    setSearchQuery(event.target.value);
-    // Auto-expand groups that have matching items
-    if (event.target.value) {
-      const lowercasedQuery = event.target.value.toLowerCase();
-      const matchingGroups = HELP_GROUPS.filter((groupTitle) => {
-        const groupItems = getGroupItems(groupTitle);
-        return groupItems.some(item =>
-          item.description.toLowerCase().includes(lowercasedQuery)
-          || item.key.toLowerCase().includes(lowercasedQuery),
-        );
-      });
-      if (matchingGroups.length > 0) {
-        setExpandedGroup(matchingGroups[0]);
-      }
-    } else {
-      setExpandedGroup(null);
-    }
-  };
-
-  // Check if any groups have matching items
-  const hasMatchingItems = useMemo(() => {
-    if (!searchQuery)
-      return true;
-    return HELP_GROUPS.some((groupTitle) => {
-      const groupItems = getGroupItems(groupTitle);
-      return groupItems.some(item =>
-        item.description.toLowerCase().includes(searchQuery.toLowerCase())
-        || item.key.toLowerCase().includes(searchQuery.toLowerCase()),
-      );
-    });
-  }, [searchQuery, navigationItems, modeItems, autoplayItems, labelItems, generalItems]);
-
   return (
     <Dialog
       id={id}
       role="dialog"
-      aria-labelledby={titleId}
-      aria-describedby={descriptionId}
       open={true}
       onClose={handleClose}
       maxWidth="sm"
       fullWidth
       disablePortal
-      className="help-dialog"
     >
       {/* Header */}
       <Grid container component={DialogTitle}>
         <Grid size="grow">
-          <Typography id={titleId} variant="h6" fontWeight="bold">
+          <Typography variant="h6" fontWeight="bold">
             Keyboard Shortcuts
           </Typography>
         </Grid>
       </Grid>
 
       <DialogContent>
-        <Typography id={descriptionId} variant="body2" sx={{ mb: 2 }}>
-          Click on a section to view its shortcuts. Alternatively, use the search bar to find specific shortcuts.
-        </Typography>
-
-        <HelpSearch value={searchQuery} onChange={handleSearchChange} />
-
-        {!hasMatchingItems && searchQuery && (
-          <Typography
-            variant="body1"
-            color="text.secondary"
-            align="center"
-            className="help-no-results"
-            role="status"
-            aria-live="polite"
-          >
-            No results found for "
-            {searchQuery}
-            "
-          </Typography>
-        )}
-
-        <Grid
-          container
-          spacing={2}
-          id="help-groups"
-          role="region"
-          aria-label="Keyboard shortcuts groups"
-          className="help-group-list"
-        >
-          {HELP_GROUPS.map(groupTitle => (
-            <Grid key={groupTitle} size={12} className="help-group">
-              <HelpGroup
-                title={groupTitle}
-                items={getGroupItems(groupTitle)}
-                isExpanded={expandedGroup === groupTitle}
-                onToggle={() => handleGroupToggle(groupTitle)}
-                searchQuery={searchQuery}
-              />
-            </Grid>
+        <Grid container spacing={1}>
+          {items.map((item, index) => (
+            <React.Fragment key={index}>
+              <Grid size={12} key={index}>
+                <HelpRow
+                  label={item.description}
+                  shortcut={item.key}
+                />
+              </Grid>
+              {index !== items.length - 1 && (
+                <Grid size={12}>
+                  <Divider />
+                </Grid>
+              )}
+            </React.Fragment>
           ))}
         </Grid>
       </DialogContent>
@@ -183,12 +94,7 @@ const Help: React.FC = () => {
           sx={{ px: 2, py: 1 }}
         >
           <Grid size="auto">
-            <Button
-              variant="contained"
-              color="primary"
-              onClick={handleClose}
-              aria-label="Close keyboard shortcuts dialog"
-            >
+            <Button variant="contained" color="primary" onClick={handleClose}>
               Close
             </Button>
           </Grid>
